lab2/requirements.txt
###### Requirements without Version Specifiers ######`
###### pip3 install -r requirements.txt        ######`
et-xmlfile
openpyxl
pyyaml
